+++ src/scalecast/SeriesTransformer.py
@@ -117,7 +117,6 @@ class SeriesTransformer:
                     "TestSetPredictions",
                     "TestSetUpperCI",
                     "TestSetLowerCI",
-                    "CIPlusMinus",
                     "UpperCI",
                     "LowerCI",
                     "TestSetActuals",
@@ -618,7 +617,6 @@ class SeriesTransformer:
                 seasrevert(h['FittedVals'],self.f.y.values[-len(h['FittedVals'])-m:], m)
             )[m:]
             h['LevelFittedVals'] = h['FittedVals'][:]
-            #ci_range = self.f._find_cis(self.f.y.values[-len(h['FittedVals']):],h['FittedVals'])
             for k in ("LevelUpperCI","UpperCI","LevelLowerCI","LowerCI"):
                 h[k] = list(seasrevert(h[k], self.f.y[-m:], m))[m:]
             for k in ("TestSetUpperCI","LevelTSUpperCI","TestSetLowerCI","LevelTSLowerCI"):